web: finish rename of ErrorPane to AlertPane
