+++ web/src/AlertPane.scss
@@ -1,6 +1,6 @@
 @import "constants";
 
-.ErrorPane {
+.AlertPane {
   margin-top: $tabnav-height;
   margin-right: $sidebar-width;
   margin-bottom: $statusbar-height;
@@ -10,12 +10,12 @@
   justify-content: center;
 }
 
-.ErrorPane ul {
+.AlertPane ul {
   list-style: none;
   width: 100%;
 }
 
-.ErrorPane-item header {
+.AlertPane-item header {
   background-color: $color-gray;
   border: 1px solid $color-gray-light;
   display: flex;
@@ -26,7 +26,7 @@
   padding-right: $spacing-unit / 2;
 }
 
-.ErrorPane-item section {
+.AlertPane-item section {
   background-color: $color-gray-dark;
   border-bottom-left-radius: $spacing-unit / 4;
   border-bottom-right-radius: $spacing-unit / 4;
@@ -36,12 +36,12 @@
   padding: $spacing-unit / 2;
 }
 
-.ErrorPane-item + .ErrorPane-item {
+.AlertPane-item + .AlertPane-item {
   margin-top: $spacing-unit;
 }
 
 // Hide all lines by default
-.ErrorPane-item code {
+.AlertPane-item code {
   display: block;
   white-space: pre-wrap;
   height: 0;
@@ -49,7 +49,7 @@
 }
 
 // ...only show last few lines
-.ErrorPane-item code:nth-last-child(-n + 8) {
+.AlertPane-item code:nth-last-child(-n + 8) {
   min-height: $spacing-unit / 2; // Respect blank lines
   height: auto;
   opacity: 1;
