+++ web/src/HUD.tsx
@@ -15,7 +15,7 @@ import { incr, pathToTag } from "./analytics"
 import TopBar from "./TopBar"
 import "./HUD.scss"
 import { TiltBuild, ResourceView } from "./types"
-import AlertPane, { AlertResource } from "./ErrorPane"
+import AlertPane, { AlertResource } from "./AlertPane"
 import PreviewList from "./PreviewList"
 import { HotKeys } from "react-hotkeys"
 import AnalyticsNudge from "./AnalyticsNudge"
